third_party/WebKit/LayoutTests/http/tests/credentialmanager/credentialscontainer-get-warnings.html

<!DOCTYPE html>
<title>Credential Management API: get() warnings.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
test(function (t) {
    navigator.credentials.get({
        unmediated: true,
        mediation: "optional"
    });
}, "mediation: 'optional' overrides unmediated: true");

test(function (t) {
    navigator.credentials.get({
        unmediated: true,
        mediation: "required"
    });
}, "mediation: 'required' overrides unmediated: true");

test(function (t) {
    navigator.credentials.get({
        unmediated: false,
        mediation: "silent"
    });
}, "mediation: 'silent' overrides unmediated: false");

test(function (t) {
    navigator.credentials.get({
        unmediated: false,
        mediation: "required"
    });
}, "mediation: 'required' overrides unmediated: false");
</script>
