third_party/WebKit/LayoutTests/http/tests/credentialmanager/credentialscontainer-get-warnings-expected.txt

CONSOLE WARNING: line 7: The boolean flag CredentialRequestOptions.unmediated is deprecated and will be removed in M62, around October 2017. Please use the CredentialRequestOptions.mediation enum instead. See https://www.chromestatus.com/features/6076479909658624 for more details.
CONSOLE WARNING: line 7: mediation: 'optional' overrides unmediated: true.
CONSOLE WARNING: line 14: mediation: 'required' overrides unmediated: true.
CONSOLE WARNING: line 21: mediation: 'silent' overrides unmediated: false.
CONSOLE WARNING: line 28: mediation: 'required' overrides unmediated: false.
This is a testharness.js-based test.
PASS mediation: 'optional' overrides unmediated: true
PASS mediation: 'required' overrides unmediated: true
PASS mediation: 'silent' overrides unmediated: false
PASS mediation: 'required' overrides unmediated: false
Harness: the test ran to completion.
